This commit is another large one, see full log
The only significant change here is the adding of Function::passive to get
ahold of all the lines. This is currently only used in markov, to add the
things to the markov model. The rest of the changes are debug info which will
be removed, or getting compile warnings to go away