Huge reworking of markov module which is probably pointless, see full log
The purpose of this change is to use vector's inside the markov model instead
of a map to a map. This was to attempt to reduce memory usage, as previously
there was a large amount of space that could be saved. There is still a large
amount of space to be saved. So far, this single change has led to roughly a
50% reduce in memory usage.

Because we're also moving towards lower order model smoothing via linear
interpolation of the probabilities of the next word, some changes to the
actual model were made as well. We now also keep around 0th-order data, in the
special empty string prefix. This is probably a large contributer to the
newfound slowness in the markov model, as this vector is consulted for every
fetch operation. Since the 0th order model is a rather special case, this will
probably be turned back into a map, but it's own independant one.

Another change to the model is that the special end of string marker is no
longer stored. Instead, the probablity of ending has been reworked so that if
we end with a punctuation mark we are more likely to stop, otherwise we end up
with longer chains on average from before. This data may be added back in
later.