Start more advance interpreted math language, see full log
The variable map was moved into the math module. The starts of a parser are in
there. I was winging it so it's really bad. Also, it only has like 45mins of
work, so... yeah. Probably going to get scratched.

GMP is now employed to get arbitrary precision ints/floats. There is a
Variable class which is moronic. So many things that were done rashly to hack
this together that I'm now commiting :)