Major improvements in parser, see full log
There has been a bunch of cleanup like converting certain functions into
member functions.

No expression has an edge with another expression, however one expression can
be completely nested inside of another. That simplifies a bunch of code.

Currently, function calling does not work as expected and there are a few
special cases that it would be nice to eliminate.