Major reworking to fix subexpression usage; see full log
Essentially, the problem was that subexpressions were not treated properly
when there was more than one statement in them. So, now when one is hit in
treeify a special node is created for it to ensure the grouping isn't lost.
This also means that toString is simpler and outputs way less parenthesis now
since we're no longer throwing away that piece of information.

In addition, function binding was made absolute lowest priority. This means it
now errors if you try to assign a function binding into a variable. Which I
guess is good :p