Start switching to new module setup; see full log
NOTICE: the repository is currently uncompilable. If I had made the push to
get it compilable in this commit, it would have been huge and unpretty.

Currently the only left modules/* files are markov.?pp. These use the new
format for module and function definition. A parser will come along for this
format soon, which will generate the needed functions inside of lib/modules.

src/global is currently based on the old system of evaluation. A lot of code
there will be deleted soon and replaced with calls into the new scripting
system.