Grab function body before clearing $# vars
This re-enables the usage of
	:eval => !1
as a simple "eval argument" function. It originally worked because 1 was
passed from the outer to inner context unchanged. When the arguments were
properly cleared and reset when entering/exiting a function, this had broke.
Grabbing the function body at the start means that if we're evaluating an
argument, everything works hunky-dory.