Rock the cache-bah. Make ngram reuse all the things.
This removes the need to recreate all the prepared statements used by the
ngramStore. Most statements are used many many many times, especially if we
limit the total order of the model.

This caches the status of tables, if they're created then we never run another
'create table if not exists'. It caches the constructed string-sql. And
finally it caches and keeps around the prepared statements made out of said
strings.