Large commit; gut brain; see full log
This is a changeset that got a little out of hand. Initially the work was
simply to switch Dictionary over to being sqlite3 backed, but that drug in the
variable store (which is currently broken).

Since these were the last things actually depending on the brain code, the
brain code has been deleted.

That caused some changes in numerous other files to remove trailing references
to brain files.

makemodes was altered, since currently no module has a save or load function
it does not support them. In the future it may be nice to have module
init/deinit functions.