Rewrite journal to be sqlite backed
This also changes the interface to the way we search the journal. Instead of
having a regex based search with an optional predicate, all searches are
predicate based. A few helpers to create a regex-based predicate and to
combine two predicates are given.

Some longstanding issues with the journal have also been resolved, such as the
output from pbrane preceding the input causing the output.